Moved the enums into separate files
