+++ src/devices/Ad7193/Ad7193.cs
@@ -38,73 +38,6 @@
         public BlockingCollection<AdcValue> AdcValues = new BlockingCollection<AdcValue>();
         public event EventHandler<AdcValueReceivedEventArgs> AdcValueReceived;
 
-        /// <summary>
-        /// AD7193 Register Map
-        /// </summary>
-        protected enum Register : byte
-        {
-            Communications = 0,     // Communications Register (WO, 8-bit)
-            Status = 0,             // Status Register         (RO, 8-bit)
-            Mode = 1,               // Mode Register           (RW, 24-bit)
-            Configuration = 2,      // Configuration Register  (RW, 24-bit)
-            Data = 3,               // Data Register           (RO, 24/32-bit)
-            ID = 4,                 // ID Register             (RO, 8-bit)
-            GPOCON = 5,             // GPOCON Register         (RW, 8-bit)
-            Offset = 6,             // Offset Register         (RW, 24-bit)
-            FullScale = 7,          // Full-Scale Register     (RW, 24-bit)
-        }
-
-        /// <summary>
-        /// Communications Register Bit Designations
-        /// </summary>
-        protected enum CommunicationsRegisterBits : byte
-        {
-            WriteEnable = (1 << 7),
-            WriteOperation = (0 << 6),
-            ReadOperation = (1 << 6),
-            ContinuousDataRead = (1 << 2)
-        }
-
-        // Gain settings
-        public enum Gain
-        {
-            X1 = 0b000,
-            X8 = 0b011,
-            X16 = 0b100,
-            X32 = 0b101,
-            X64 = 0b110,
-            X128 = 0b111
-        }
-
-        [Flags]
-        public enum Channel
-        {
-            CH00 = 0b00_0000_0001,
-            CH01 = 0b00_0000_0010,
-            CH02 = 0b00_0000_0100,
-            CH03 = 0b00_0000_1000,
-            CH04 = 0b00_0001_0000,
-            CH05 = 0b00_0010_0000,
-            CH06 = 0b00_0100_0000,
-            CH07 = 0b00_1000_0000,
-            TEMP = 0b01_0000_0000,
-            Shrt = 0b10_0000_0000
-        }
-
-        public enum AveragingMode
-        {
-            Off = 0b00,
-            Avg2 = 0b01,
-            Avg8 = 0b10,
-            Avg16 = 0b11
-        }
-
-        public enum AnalogInputMode
-        {
-            FourDifferentialAnalogInputs = 0b0,
-            EightPseudoDifferentialAnalogInputs = 0b1
-        }
-
         // Default register settings
         protected uint[] registerCache = { 0x00, 0x080060, 0x000117, 0x000000, 0xa2, 0x00, 0x000000, 0x000000 };
         protected byte[] registerSize = { 1, 3, 3, 3, 1, 1, 3, 3 };
