+++ src/devices/Ad7193/AnalogInputMode.cs
@@ -1,0 +1,8 @@
+﻿namespace Iot.Device.Ad7193
+{
+    public enum AnalogInputMode
+    {
+        FourDifferentialAnalogInputs = 0b0,
+        EightPseudoDifferentialAnalogInputs = 0b1
+    }
+}
