+++ src/devices/Ad7193/AveragingMode.cs
@@ -1,0 +1,10 @@
+﻿namespace Iot.Device.Ad7193
+{
+    public enum AveragingMode
+    {
+        Off = 0b00,
+        Avg2 = 0b01,
+        Avg8 = 0b10,
+        Avg16 = 0b11
+    }
+}
