+++ src/devices/Ad7193/Gain.cs
@@ -1,0 +1,13 @@
+﻿namespace Iot.Device.Ad7193
+{
+    // Gain settings
+    public enum Gain
+    {
+        X1 = 0b000,
+        X8 = 0b011,
+        X16 = 0b100,
+        X32 = 0b101,
+        X64 = 0b110,
+        X128 = 0b111
+    }
+}
